+++ mtenn/representation.py
@@ -1,0 +1,5 @@
+import torch
+
+
+class Representation(torch.nn.Module):
+    pass
